requirements.txt
pyzmq
numpy
pymap3d
scipy
websockets
